backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
firebase-admin==6.5.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-firestore==2.19.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
resend==2.4.0
